fix(daemon): shutdown-aware background tasks (handoff item 10)
Live 2026-07-08: SIGTERM breached uvicorn's graceful window ("Cancel 54
running task(s), timeout graceful shutdown exceeded" + CancelledError
tracebacks from in-flight ASGI stacks). Three structural causes:

(a) Fire-and-forget asyncio.create_task call sites in routes.py (usage
    bumps, demand logging, HyDE/rerank cache warms, sync trigger) were
    untracked -- nothing owned them at shutdown.
(b) Background passes self-HTTP the daemon's own endpoints
    (_fetch_type_map, _fetch_tool_rule_ids, hygiene's list/delete calls).
    Once uvicorn stops serving, those requests can never complete,
    guaranteeing the timeout breach if a pass is mid-flight.
(c) asyncio.to_thread workers blocked on the process-global LLAMA lock are
    uncancellable -- cancellation must be bounded-wait-then-abandon.

New simba.memory.background module: spawn() creates+tracks a task with
auto-discard on completion; drain(grace_seconds) waits for tasks, cancels
stragglers, and abandons anything still running after a short bounded tail
rather than hanging forever. A process-level shutdown flag
(mark_shutting_down/is_shutting_down, backed by threading.Event so it's
safe to read from to_thread workers) lets self-HTTP helpers bail out
immediately instead of attempting a loopback call that can never complete.

Converted all 7 fire-and-forget create_task call sites in routes.py to
background.spawn(); routes.py's existing _background_tasks set is now an
alias for background.TASKS so hybrid.py's bg_tasks= reranker task is
covered by drain() too, with no changes to hybrid.py itself. Guarded the
two named self-HTTP helpers in maintenance.py plus hygiene.py's
run_hygiene_pass (both its initial list fetch and its per-item delete
loop) -- the other default-on MaintenanceScheduler self-HTTP surface.
decay.py's capacity-cap fetch and the episodes/reflection/sync-scheduler
self-calls are left unguarded: all are gated behind default-off config
levers, so they're not part of the live daemon's default background
behavior today.

Reused the existing memory.shutdown_timeout knob (added Feb 9 for the
sync-task shutdown dance and uvicorn's timeout_graceful_shutdown) instead
of adding a new one -- widened int->float so tests can drive a sub-second
grace without truncation. It now also feeds background.drain()'s budget in
server.py's shutdown hook, extracted into _shutdown_daemon() so it's
testable without a real DB/embedding-model startup. mark_shutting_down()
runs first, before scheduler.stop(), so a pass already mid-flight sees the
flag on its very next self-HTTP attempt.

MaintenanceScheduler.stop() was already invoked at shutdown and
run_forever() already avoids clearing a pre-set stop event -- both
verified unchanged.

+++ src/simba/memory/config.py
@@ -130,7 +130,16 @@ class MemoryConfig:
     # Let the background sync scheduler run the hygiene pass.
     hygiene_scheduler_enabled: bool = True
     sync_interval: int = 0
-    shutdown_timeout: int = 10
+    # Graceful-shutdown budget (seconds) --- the ONE knob governing both
+    # uvicorn's own `timeout_graceful_shutdown` (server.py's uvicorn.run
+    # call) and `simba.memory.background.drain()`'s wait-for-fire-and-forget-
+    # tasks budget in the daemon's shutdown hook (handoff item 10, spec 33
+    # follow-up). Float (was int prior to this change) so tests can drive a
+    # sub-second grace without truncation. Live 2026-07-08: a SIGTERM
+    # breached this window ("Cancel 54 running task(s)") because those
+    # fire-and-forget tasks were never tracked/awaited at all; see
+    # background.py's module docstring for the full failure analysis.
+    shutdown_timeout: float = 10.0
     # Continuous extraction (the "Continuous" gap). When on, the Stop hook reads only
     # the NEW transcript window each turn (incremental cursor, O(new)) and enqueues it
     # for the scored extract->score->keep/drop worker. DEFAULT-OFF: this ships the
